fixed: very short wavs (< 1 second) could not be played (PAPlayer) fixed: mono mp3's could not be played (PAPlayer, MadCodec.dll)
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@4382 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/paplayer/DTSCodec.cpp
@@ -269,7 +269,7 @@ bool DTSCodec::CalculateTotalTime()
   //instead of this one
   //m_TotalTime = (((m_file.GetLength() - m_iDataStart) * 8) / m_Bitrate) * 1000;
   m_TotalTime = (m_file.GetLength() - m_iDataStart) / (m_SampleRate * 2 * (m_BitsPerSample / 8)) * 1000;
-  m_Bitrate = (int)(((m_file.GetLength() - m_iDataStart) * 8) / (m_TotalTime / 1000));
+  m_Bitrate = (int)(((m_file.GetLength() - m_iDataStart) * 8) / float(m_TotalTime / 1000));
   return m_TotalTime > 0;
 }
